parse buffer from different bases and data unit: (#754)
All flags that expect buffer as input i.e. `--output-file-size-limit`, `--output-file-max-size-limit`, `--copy-buffer-size` and `input-raw-buffer-size` can now parse  inputs from differents bases and data units like: `10mb`, `10kb`, `100gb`,   `18tb`, `11839023`.... 
data units and bases are case insensitive, the parser accepts only the format of [Go integer literals](https://golang.org/ref/spec#Integer_literals)

+++ input_raw.go
@@ -21,7 +21,7 @@ type RAWInput struct {
 	listener      *raw.Listener
 	bpfFilter     string
 	timestampType string
-	bufferSize    int
+	bufferSize    int64
 }
 
 // Available engines for intercepting traffic
@@ -32,7 +32,7 @@ const (
 )
 
 // NewRAWInput constructor for RAWInput. Accepts address with port as argument.
-func NewRAWInput(address string, engine int, trackResponse bool, expire time.Duration, realIPHeader string, bpfFilter string, timestampType string, bufferSize int) (i *RAWInput) {
+func NewRAWInput(address string, engine int, trackResponse bool, expire time.Duration, realIPHeader string, bpfFilter string, timestampType string, bufferSize int64) (i *RAWInput) {
 	i = new(RAWInput)
 	i.data = make(chan *raw.TCPMessage)
 	i.address = address
@@ -105,6 +105,7 @@ func (i *RAWInput) String() string {
 	return "Intercepting traffic from: " + i.address
 }
 
+// Close closes the input raw listener
 func (i *RAWInput) Close() error {
 	i.listener.Close()
 	close(i.quit)
